Abstract new package: signing
Move signing related code into new package
