+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/server/signing/SigningControllerConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.credentials;
+package io.trino.s3.proxy.server.signing;
 
 import io.airlift.configuration.Config;
 import io.airlift.units.Duration;
